feat - teste unitario delete

+++ app/Repository/Eloquent/UserRepository.php
@@ -4,6 +4,9 @@ namespace App\Repository\Eloquent;
 
 use App\Models\User;
 use App\Repository\Contracts\UserRepositoryInterface;
+use App\Repository\Exceptions\NotFoundException;
+use Exception;
+use PhpParser\Node\Expr\Throw_;
 
 class UserRepository implements UserRepositoryInterface
 {
@@ -39,7 +42,9 @@ class UserRepository implements UserRepositoryInterface
     public function delete(string $email): bool
     {
         $user = $this->model->where('email', $email)->first();
-
+        if(!$user){
+            throw new NotFoundException("User not found");
+        }
         return $user->delete();
